add fetch encoded update
